online_learning/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==3.2.12
django-appconf==1.0.5
django-braces==1.15.0
django-cors-headers==3.11.0
django-daterange-filter==1.3.0
django-filter==21.1
django-imagekit==4.1.0
django-import-export==2.7.1
django-model-utils==4.2.0
django-oauth-toolkit==1.7.0
django-qiniu-storage==2.3.1
django-rest-framework-social-oauth2==1.1.0
django-simple-sso==1.1.0
django-simpleui==2022.2
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
flake8==4.0.1
idna==3.3
importlib-metadata==4.2.0
inflection==0.5.1
itsdangerous==0.24
itypes==1.2.0
Jinja2==3.0.3
jwcrypto==1.0
MarkupPy==1.14
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pilkit==2.0
Pillow==9.0.1
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.11.2
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
qiniu==7.5.0
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==3.4.0
sqlparse==0.4.2
tablib==3.2.0
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
webservices==0.7
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
yapf==0.32.0
zipp==3.7.0
